girepository: Fix -Wsign-conversion warnings with string arithmetic
There are a few `g_strndup()` calls which use a length calculated from
the return value of `strchr()` minus the original string. That’s fine,
as long as `strchr()` doesn’t return `NULL`. Add some asserts to ensure
that.


Helps: #3405

+++ girepository/girparser.c
@@ -707,7 +707,7 @@ parse_type_internal (GIIrModule   *module,
              *str == ':')
         (str)++;
 
-      type->giinterface = g_strndup (start, str - start);
+      type->giinterface = g_strndup (start, (size_t) (str - start));
     }
 
   if (next)
